file_encryption/src/requirements.txt
altgraph==0.17
future==0.18.2
pefile==2019.4.18
pyasn1==0.4.8
pycryptodome==3.9.7
pycryptodomex==3.9.7
PyInstaller==3.6
PyQt5==5.14.2
PyQt5-sip==12.7.2
PyQt5-stubs==5.14.2.0
pywin32-ctypes==0.2.0
rsa==4.0
